pyproject.toml
[tool.poetry]
name = "spotify_app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
apache-airflow = "2.3.0"
requests = "^2.27.1"
pandas = "^1.4.2"
spotipy = "^2.19.0"
alembic = "^1.7.7"
paramiko = "^2.11.0"
pendulum = "^2.1.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
ipdb = "^0.13.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# @TODO investigate this
# [tool.poetry.scripts]
# my-script = 'dags:tutorial_dag.py'
